Finish tournament if lack of players and fix bugs

+++ tg/tournament/announcements.py
@@ -12,6 +12,7 @@ from tg.utils import (
     get_tournament_welcome_message,
     Button,
     get_next_tour_message,
+    get_tournament_end_without_players_message,
 )
 
 
@@ -58,9 +59,24 @@ def announce_tournament_end(tournament_db: TournamentDB, bot: TeleBot):
     tournament_url = tournament_db.get_url()
     results = tournament_db.get_final_results()
     winners = [player.tg_username.value for player in results][:3]
-    bot.delete_message(chat_id, settings.registration_list_message_id.value)
+    if settings.registration_list_message_id.value:
+        bot.delete_message(chat_id, settings.registration_list_message_id.value)
     bot.send_message(
         chat_id,
         get_tournament_end_message(winners, tournament_url),
+        message_thread_id=message_thread_id,
+    )
+
+
+def announce_tournament_end_without_players(tournament_db: TournamentDB, bot: TeleBot):
+    nmd_logger.info("End tournament announcement")
+    chat_id = settings_db.settings.chat_id.value
+    message_thread_id = settings_db.settings.tournament_thread_id.value
+    settings = tournament_db.settings
+    if settings.registration_list_message_id.value:
+        bot.delete_message(chat_id, settings.registration_list_message_id.value)
+    bot.send_message(
+        chat_id,
+        get_tournament_end_without_players_message(),
         message_thread_id=message_thread_id,
     )
